Cart payment duyloc

+++ src/java/com/cellphone/controller/HomeServlet.java
@@ -46,10 +46,21 @@ public class HomeServlet extends HttpServlet {
                 product.setColor(result.getString("typeByColor"));
                 product.setImage(result.getString("image"));
                 product.setDesc(result.getString("description"));
-                product.setPrice(result.getInt("price"));
+                product.setPrice(result.getDouble("price"));
                 product.setSlug(result.getString("slug"));
                 
                 productList.addProduct(product);
+//                System.out.println("Số lượng sản phẩm đã lấy: " + productList.getProductList().size());
+//                
+//                System.out.println("Sản phẩm đã lấy: ");
+//                System.out.println("ID: " + product.getId());
+//                System.out.println("Tên: " + product.getName());
+//                System.out.println("Hãng: " + product.getBrand());
+//                System.out.println("Màu: " + product.getTypeByColor());
+//                System.out.println("Hình ảnh: " + product.getImage());
+//                System.out.println("Mô tả: " + product.getDescription());
+//                System.out.println("Giá: " + product.getPrice());
+//                System.out.println("Slug: " + product.getSlug());
             }
         } catch (SQLException e) {
             throw new ServletException("Lỗi kết nối CSDL: " + e.getMessage(), e);
